Fixes #2388 MoBi.R back to honest cross-platform net10.0
Removes OSPSuite.Presentation from MoBi.R's transitive closure so the
assembly can load on Linux / macOS .NET 10. Result:

  MoBi.R, MoBi.R.Tests, MoBi.CLI.Core, MoBi.CLI -> net10.0
  MoBi.Core                                     -> net10.0 (no System.Drawing.Common)
  MoBi.Presentation, MoBi.UI                    -> net10.0-windows (unchanged)

  dotnet nuget why src/Mobi.R/MoBi.R.csproj OSPSuite.Presentation -> no path
  2509 MoBi.Tests pass, 224 MoBi.R.Tests pass

Per-class summary:

- IDiagramModelToImageTask: moved up from MoBi.Core to MoBi.Presentation.
  Returns System.Drawing.Bitmap (Windows-only via System.Drawing.Common).
  After the move MoBi.Core drops its explicit System.Drawing.Common
  PackageReference.

- ISerializationTask / SerializationTask: split. New
  MoBi.Core.Serialization.Services.ICoreSerializationTask /
  CoreSerializationTask in MoBi.Core (8 collaborators, no LoadMany, no
  IXmlContentSelector; Load<T> is LoadAll<T>(file).FirstOrDefault()).
  Presentation's SerializationTask now inherits from CoreSerializationTask
  and adds LoadMany<T> plus the IXmlContentSelector constructor parameter.
  LoadMany stays in Presentation because it disambiguates multi-match XML
  elements by popping a modal dialog via IXmlContentSelector /
  IModalPresenter, which has no headless equivalent. R only ever calls
  LoadAll<T>.

- MoBiXmlSerializerRepository: split via inheritance. New
  MoBi.Core.Serialization.Xml.Serializer.MoBiCoreXmlSerializerRepository
  does base.AddInitialSerializer(), the three RemoveSerializer calls for
  the OSPSuite types MoBi overrides, and the MoBi.Core assembly scan for
  IOSPSuiteXmlSerializer plus the cross-platform attribute mappers
  (NodeSize, NotificationType, OriginalQuantityValue.Types). Presentation
  MoBiXmlSerializerRepository now inherits from that and adds the
  MoBi.Presentation IOSPSuiteXmlSerializer scan, the IPresentationXmlSerializer
  scan, and AddPresentationSerializers(). The Presentation half stays
  because IPresentationXmlSerializer is a Presentation contract (chart
  layouts, journal pages, UI state) and AddPresentationSerializers() lives
  in OSPSuite.Presentation.Serialization.Extensions.

- SettingsPersistor<TSettings> / ISettingsPersistor<TSettings>: straight
  move from MoBi.Presentation.Serialization.Xml to MoBi.Core.Serialization.Xml.
  The generic base only references IMoBiXmlSerializerRepository,
  IMoBiConfiguration, ISerializationContextFactory, OSPSuite.Utility
  helpers, and System.Xml.Linq.

- ApplicationSettingsPersistor: straight move to MoBi.Core alongside its
  base. Generic argument is MoBi.Core.IApplicationSettings; only Core types
  in use.

- UserSettingsPersistor: stays in MoBi.Presentation. It calls
  _userSettings.SaveLayout(), where SaveLayout() is a DevExpress dock /
  ribbon UI affordance with no headless equivalent.

- SerializerRegister: partial inline. MoBi.R/Bootstrap/ApplicationStartup
  previously called
  new SerializerRegister().PerformMappingForMoBiSerializerRepository(...).
  That method's body is two lines (resolve IMoBiXmlSerializerRepository,
  call PerformMapping()); inlined directly so R no longer compiles against
  the SerializerRegister class (which transitively references
  OSPSuite.Presentation.Serialization.Extensions). The class itself stays
  in Presentation for desktop callers.

- MoBiApplicationController: stays in MoBi.Presentation. Inherits
  OSPSuite.Presentation.Core.ApplicationController, returns IModalPresenter,
  and switches on building-block subtypes to start IEdit*Presenter
  dialogs. R no longer registers it; the only reason it was registered
  before was that XmlContentSelector transitively depended on it.

- XmlContentSelector: stays in MoBi.Presentation. Pops a modal dialog
  (IModalPresenter / ISelectManyPresenter<XElement>). R no longer registers
  it because the new headless CoreSerializationTask doesn't accept
  IXmlContentSelector at all.

- CoreUserSettings (new): minimal stub in
  MoBi.CLI.Core.MinimalImplementations. Replaces the old four-interface
  UserSettings stub. Satisfies only MoBi.Core.ICoreUserSettings and
  OSPSuite.Core.ICoreUserSettings (what MoBi.Core.DimensionValidator
  actually needs) -- no IPresentationUserSettings, no IUserSettings.

- MoBi.CLI.Core/MinimalImplementations/UserSettings.cs: deleted. The old
  stub implemented MoBi.Presentation.Settings.IUserSettings (which extends
  OSPSuite.Presentation.IPresentationUserSettings), which is what dragged
  OSPSuite.Presentation into MoBi.CLI.Core.

- MoBi.CLI.Core.ApplicationStartup: rewired. Uses
  OSPSuite.Core.Serialization.CoreSerializerRegister directly instead of
  the Presentation SerializerRegister, registers MoBiCoreXmlSerializerRepository
  (Core variant) for IMoBiXmlSerializerRepository, registers the new
  CoreUserSettings stub, and inlines the serializer-mapping call. After
  this, MoBi.CLI.Core has zero Presentation references.

- MoBi.R/RRegister.cs: aggressively stripped. Removed defensive
  registrations of IMoBiApplicationController, IXmlContentSelector,
  the four-interface UserSettings, IDiagramModelToXmlMapper,
  IJournalDiagramManagerFactory, and the diagram-manager factory bindings
  that no R code path actually resolves. Restored only what the build /
  tests demanded: the three diagram-manager stubs (consumed by Core XML
  serializers via serializationContext.Resolve<>), CoreUserSettings
  (consumed by DimensionValidator), CoreSerializationTask, the Core XML
  serializer repository, and the basic Core persistence services.

- Convention-collision rename: the OSPSuiteRegistrationConvention auto-
  registers class X against IX by simple name. The split classes were
  initially named the same on both sides (MoBiXmlSerializerRepository in
  both Core and Presentation; SerializationTask in both) and the convention
  registered both against IMoBiXmlSerializerRepository /
  IOSPSuiteXmlSerializerRepository / ISerializationTask, causing 24
  integration-test failures from non-deterministic resolution. Renamed the
  Core variants to MoBiCoreXmlSerializerRepository, CoreSerializationTask,
  and ICoreSerializationTask to remove the simple-name collision entirely.

- csproj flips. MoBi.R, MoBi.R.Tests, MoBi.CLI.Core, MoBi.CLI all moved
  from net10.0-windows to net10.0. MoBi.CLI also dropped
  <UseWindowsForms>True</UseWindowsForms> (dead flag; no WinForms in source)
  and the MoBi.Presentation ProjectReference. MoBi.Core dropped the
  explicit System.Drawing.Common PackageReference.
