+++ src/MoBi.CLI.Core/ApplicationStartup.cs
@@ -1,32 +1,31 @@
-﻿using Castle.Facilities.TypedFactory;
+using System.Threading;
+using Castle.Facilities.TypedFactory;
 using MoBi.Assets;
 using MoBi.CLI.Core.MinimalImplementations;
 using MoBi.Core;
 using MoBi.Core.Domain.Model.Diagram;
 using MoBi.Core.Extensions;
+using MoBi.Core.Serialization.Xml;
+using MoBi.Core.Serialization.Xml.Serializer;
 using MoBi.Core.Serialization.Xml.Services;
-using MoBi.Presentation.Serialization;
-using MoBi.Presentation.Serialization.Xml;
-using MoBi.Presentation.Serialization.Xml.Serializer;
-using MoBi.Presentation.Settings;
 using OSPSuite.Core;
 using OSPSuite.Core.Domain;
 using OSPSuite.Core.Domain.Builder;
 using OSPSuite.Core.Domain.UnitSystem;
+using OSPSuite.Core.Serialization;
 using OSPSuite.Core.Serialization.Xml;
 using OSPSuite.Infrastructure;
 using OSPSuite.Infrastructure.Container.Castle;
-using OSPSuite.Presentation;
 using OSPSuite.Utility.Container;
 using OSPSuite.Utility.Events;
 using OSPSuite.Utility.Extensions;
 using OSPSuite.Utility.FileLocker;
 using OSPSuite.Utility.Format;
-using System.Threading;
 using ApplicationSettings = MoBi.Core.ApplicationSettings;
 using CoreRegister = MoBi.Core.CoreRegister;
 using IApplicationSettings = MoBi.Core.IApplicationSettings;
 using IContainer = OSPSuite.Utility.Container.IContainer;
+using ICoreUserSettings = MoBi.Core.ICoreUserSettings;
 
 namespace MoBi.CLI.Core
 {
@@ -50,7 +49,7 @@ namespace MoBi.CLI.Core
       {
          var container = IoC.Container;
 
-         var register = new SerializerRegister();
+         var register = new CoreSerializerRegister();
          using (container.OptimizeDependencyResolution())
          {
             container.RegisterImplementationOf(NumericFormatterOptions.Instance);
@@ -67,23 +66,19 @@ namespace MoBi.CLI.Core
 
          container.Register<IFileLocker, FileLocker>(LifeStyle.Singleton);
          container.Register<IMoBiConfiguration, IApplicationConfiguration, MoBiConfiguration>(LifeStyle.Singleton);
-         container.Register<IMoBiXmlSerializerRepository, MoBiXmlSerializerRepository>(LifeStyle.Singleton);
-         container.Register<IUserSettings, IPresentationUserSettings, OSPSuite.Core.ICoreUserSettings, MoBi.Core.ICoreUserSettings, UserSettings>(LifeStyle.Singleton);
+         container.Register<IMoBiXmlSerializerRepository, MoBiCoreXmlSerializerRepository>(LifeStyle.Singleton);
+         container.Register<ICoreUserSettings, OSPSuite.Core.ICoreUserSettings, CoreUserSettings>(LifeStyle.Singleton);
          container.Register<IApplicationSettings, OSPSuite.Core.IApplicationSettings, ApplicationSettings>(LifeStyle.Singleton);
          container.Register<ISpatialStructureDiagramManager, SpatialStructureDiagramManager>();
          container.Register<IMoBiReactionDiagramManager, MoBiReactionDiagramManager>();
          container.Register<ISimulationDiagramManager, SimulationDiagramManager>();
-         container.Register<ISettingsPersistor<IUserSettings>, UserSettingsPersistor>();
          container.Register<ISettingsPersistor<IApplicationSettings>, ApplicationSettingsPersistor>();
 
-         
-
          register.PerformMappingForSerializerIn(container);
+         container.Resolve<IMoBiXmlSerializerRepository>().PerformMapping();
+
          initializeDimensions(container);
          initCalculationMethodRepository(container);
-
-         var userSettingsPersistor = container.Resolve<ISettingsPersistor<IUserSettings>>();
-         userSettingsPersistor.Load();
 
          var applicationSettingsPersistor = container.Resolve<ISettingsPersistor<IApplicationSettings>>();
          applicationSettingsPersistor.Load();
