+++ src/MoBi.CLI.Core/MoBi.CLI.Core.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-	  <TargetFramework>net10.0-windows</TargetFramework>
+	  <TargetFramework>net10.0</TargetFramework>
       <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
       <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
       <PackageTags>open-systems-pharmacology, ospsuite-components</PackageTags>
@@ -16,18 +16,16 @@
 
 
 	<ItemGroup>
-      <PackageReference Include="OSPSuite.Assets" Version="13.0.111" />
-      <PackageReference Include="OSPSuite.Infrastructure.Autofac" Version="13.0.111" />
+      <PackageReference Include="OSPSuite.Assets" Version="13.0.113" />
+      <PackageReference Include="OSPSuite.Infrastructure.Autofac" Version="13.0.113" />
       <PackageReference Include="OSPSuite.Utility" Version="5.0.0.7" />
-      <PackageReference Include="OSPSuite.Core" Version="13.0.111" />
-      <PackageReference Include="OSPSuite.Presentation" Version="13.0.111" />
-      <PackageReference Include="OSPSuite.R" Version="13.0.111" />
-      <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.111" />
-      <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.111" />
+      <PackageReference Include="OSPSuite.Core" Version="13.0.113" />
+      <PackageReference Include="OSPSuite.R" Version="13.0.113" />
+      <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.113" />
+      <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.113" />
 	</ItemGroup>
 	<ItemGroup>
       <ProjectReference Include="..\MoBi.Assets\MoBi.Assets.csproj" />
-      <ProjectReference Include="..\MoBi.Presentation\MoBi.Presentation.csproj" />
       <ProjectReference Include="..\MoBi.Core\MoBi.Core.csproj" />
 	</ItemGroup>
    <ItemGroup>
