+++ src/MoBi.CLI/MoBi.CLI.csproj
@@ -2,8 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <UseWindowsForms>True</UseWindowsForms>
-    <TargetFramework>net10.0-windows</TargetFramework>
+    <TargetFramework>net10.0</TargetFramework>
     <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
     <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
     <Authors>Open-Systems-Pharmacology</Authors>
@@ -22,25 +21,21 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="9.0.6" />
     <PackageReference Include="CommandLineParser" Version="2.9.1" />
-    <PackageReference Include="OSPSuite.Core" Version="13.0.111" />
-    <PackageReference Include="OSPSuite.Presentation" Version="13.0.111" />
-    <PackageReference Include="OSPSuite.Assets" Version="13.0.111" />
+    <PackageReference Include="OSPSuite.Core" Version="13.0.113" />
+    <PackageReference Include="OSPSuite.Assets" Version="13.0.113" />
     <PackageReference Include="OSPSuite.Utility" Version="5.0.0.7" />
     <PackageReference Include="OSPSuite.FuncParser" Version="5.0.0.3" />
     <PackageReference Include="OSPSuite.SimModel" Version="5.0.0.76" />
     <PackageReference Include="OSPSuite.SimModelSolver_CVODES" Version="5.0.0.53" />
     <PackageReference Include="SixLabors.ImageSharp" Version="2.1.11" />
-    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.111" />
-    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.111" />
+    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.113" />
+    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.113" />
   </ItemGroup>
   <ItemGroup>
      <Content Include="..\..\dimensions\OSPSuite.Dimensions.xml" Link="OSPSuite.Dimensions.xml">
         <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
      </Content>
      <Content Include="..\..\pkparameters\OSPSuite.PKParameters.xml" Link="OSPSuite.PKParameters.xml">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-     </Content>
-     <Content Include="$(PkgOSPSuite_Presentation)\OSPSuite.Presentation\**">
         <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
      </Content>
      <Content Include="$(PkgOSPSuite_TeXReporting)\OSPSuite.TeXReporting\**">
