+++ src/MoBi.Core/Serialization/Services/CoreSerializationTask.cs
@@ -1,0 +1,224 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Xml.Linq;
+using MoBi.Assets;
+using MoBi.Core.Domain.Model;
+using MoBi.Core.Events;
+using MoBi.Core.Exceptions;
+using MoBi.Core.Serialization.Converter;
+using MoBi.Core.Serialization.ORM;
+using MoBi.Core.Serialization.Xml.Services;
+using OSPSuite.Core.Domain;
+using OSPSuite.Core.Domain.Services;
+using OSPSuite.Core.Serialization;
+using OSPSuite.Core.Serialization.Exchange;
+using OSPSuite.Core.Services;
+using OSPSuite.Utility.Extensions;
+
+namespace MoBi.Core.Serialization.Services
+{
+   public interface ICoreSerializationTask : ISerializationTask
+   {
+      void LoadProject(string fileName);
+      void SaveProject();
+      void CloseProject();
+      MoBiProject NewProject();
+
+      void LoadJournal(string journalPath, string projectFullPath = null, bool showJournal = false);
+
+      /// <summary>
+      ///    Loads all instances of the element type <typeparamref name="T" /> from the <paramref name="fileName" />.
+      ///    If <paramref name="resetIds" /> is true, then all element ID's will be set to new values on load.
+      /// </summary>
+      /// <returns>A new list of elements that were loaded</returns>
+      IReadOnlyList<T> LoadAll<T>(string fileName, bool resetIds = false);
+   }
+
+   public class CoreSerializationTask : ICoreSerializationTask
+   {
+      private readonly IXmlSerializationService _xmlSerializationService;
+      private readonly IContextPersistor _contextPersistor;
+      protected readonly IObjectTypeResolver _objectTypeResolver;
+      private readonly IDialogCreator _dialogCreator;
+      private readonly IProjectConverterLogger _projectConverterLogger;
+      private readonly IMoBiContext _context;
+      private readonly IPostSerializationStepsMaker _postSerializationSteps;
+      private readonly IHeavyWorkManager _heavyWorkManager;
+
+      public CoreSerializationTask(IXmlSerializationService xmlSerializationService, 
+         IContextPersistor contextPersistor,
+         IObjectTypeResolver objectTypeResolver, 
+         IDialogCreator dialogCreator,
+         IProjectConverterLogger projectConverterLogger,
+         IMoBiContext context, 
+         IPostSerializationStepsMaker postSerializationSteps, 
+         IHeavyWorkManager heavyWorkManager)
+      {
+         _xmlSerializationService = xmlSerializationService;
+         _contextPersistor = contextPersistor;
+         _objectTypeResolver = objectTypeResolver;
+         _dialogCreator = dialogCreator;
+         _projectConverterLogger = projectConverterLogger;
+         _context = context;
+         _postSerializationSteps = postSerializationSteps;
+         _heavyWorkManager = heavyWorkManager;
+      }
+
+      public void LoadProject(string fileName)
+      {
+         if (!tryLockFile(fileName))
+            return;
+
+         try
+         {
+            _heavyWorkManager.Start(() => _contextPersistor.Load(_context, fileName), AppConstants.Captions.LoadingProject);
+         }
+         catch (ConversionException e)
+         {
+            _dialogCreator.MessageBoxInfo(e.Message);
+         }
+      }
+
+      private bool tryLockFile(string projectFile)
+      {
+         try
+         {
+            _context.LockFile(projectFile);
+            return true;
+         }
+         catch (CannotLockFileException e)
+         {
+            var ans = _dialogCreator.MessageBoxYesNo(AppConstants.Exceptions.ProjectWillBeOpenedAsReadOnly(e.Message), AppConstants.Captions.OpenAnyway, AppConstants.Captions.CancelButton);
+            var shouldOpenInReadOnly = (ans == ViewResult.Yes);
+            if (!shouldOpenInReadOnly) return false;
+            _context.ProjectIsReadOnly = true;
+            return true;
+         }
+      }
+
+      public void SaveProject()
+      {
+         var project = _context.CurrentProject;
+         if (project == null) return;
+
+         _context.LockFile(project.FilePath);
+
+         _contextPersistor.Save(_context);
+
+         _context.AccessFile(project.FilePath);
+
+         _context.ProjectIsReadOnly = false;
+      }
+
+      public void CloseProject() => _contextPersistor.CloseProject(_context);
+
+      public MoBiProject NewProject() => _contextPersistor.NewProject(_context);
+
+      public void LoadJournal(string journalPath, string projectFullPath = null, bool showJournal = false)
+      {
+         if (string.IsNullOrEmpty(journalPath))
+            return;
+
+         _contextPersistor.LoadJournal(_context, journalPath, projectFullPath, showJournal);
+      }
+
+      public void SaveModelPart<T>(T entityToSerialize, string filename)
+      {
+         var xelPart = _xmlSerializationService.SerializeModelPart(entityToSerialize);
+         xelPart.PermissiveSave(filename);
+      }
+
+      public virtual T Load<T>(string fileName, bool resetIds = false) => LoadAll<T>(fileName, resetIds).FirstOrDefault();
+
+      public IReadOnlyList<T> LoadAll<T>(string fileName, bool resetIds = false) => loadMany<T>(fileName, x => x, resetIds).ToList();
+
+      protected IEnumerable<T> loadMany<T>(string fileName, Func<IReadOnlyList<XElement>, IReadOnlyList<XElement>> elementSelector, bool resetIds = false)
+      {
+         if (string.IsNullOrEmpty(fileName))
+            return Enumerable.Empty<T>();
+
+         _projectConverterLogger.Clear();
+
+         var xelRoot = XElementSerializer.PermissiveLoad(fileName);
+
+         var possibleElementsToDeserialize = retrieveElementsToDeserializeFromFile<T>(xelRoot, fileName).ToList();
+
+         var selection = possibleElementsToDeserialize.Count > 1
+            ? elementSelector(possibleElementsToDeserialize)
+            : possibleElementsToDeserialize;
+
+         if (!selection.Any())
+            return Enumerable.Empty<T>();
+
+         var deserializedObjects = new List<T>();
+         int version = _xmlSerializationService.VersionFrom(xelRoot);
+         var currentProject = _context.CurrentProject;
+         selection.Each(element => deserializedObjects.Add(_xmlSerializationService.Deserialize<T>(element, currentProject, version)));
+
+         var notificationMessages = _projectConverterLogger.AllMessages().ToList();
+         if (notificationMessages.Any())
+            _context.PublishEvent(new ShowProjectConversionNotificationsEvent(notificationMessages));
+
+         _postSerializationSteps.PerformPostDeserializationFor(deserializedObjects, version, resetIds);
+
+         return deserializedObjects;
+      }
+
+      private IEnumerable<XElement> retrieveElementsToDeserializeFromFile<T>(XElement xelRoot, string fileName)
+      {
+         var elementName = _xmlSerializationService.ElementNameFor(typeof(T));
+
+         //models only required when importing simulations. Otherwise, too many containers are loaded
+         if (!loadingSimulation<T>())
+            xelRoot = removeModelFrom(xelRoot);
+
+         var possibleElementsToDeserialize = retrieveElementsToDeserializeFromRootNode(xelRoot, elementName).ToList();
+         if (possibleElementsToDeserialize.Any())
+            return possibleElementsToDeserialize;
+
+         throw new NotMatchingSerializationFileException(fileName, _objectTypeResolver.TypeFor<T>(), xelRoot.Name.LocalName);
+      }
+
+      private IEnumerable<XElement> retrieveElementsToDeserializeFromRootNode(XElement xelRoot, string expectedElementName)
+      {
+         var possibleElementsToDeserialize = xelRoot.DescendantsAndSelf(expectedElementName).ToList();
+         if (possibleElementsToDeserialize.Any())
+            return possibleElementsToDeserialize;
+
+         //no element found. Check to see if we can match elements using a previously saved name or a build configuration
+         return xelRoot.DescendantsAndSelf(getElementNameFromBuildConfiguration(expectedElementName));
+      }
+
+      private bool loadingSimulation<T>()
+      {
+         return typeof(T).IsAnImplementationOf<IModelCoreSimulation>() ||
+                typeof(T).IsAnImplementationOf<SimulationTransfer>();
+      }
+
+      private XElement removeModelFrom(XElement xelRoot)
+      {
+         var models = xelRoot.DescendantsAndSelf(AppConstants.XmlNames.Model).ToList();
+         models.Each(model => model.Remove());
+         return xelRoot;
+      }
+
+      private string getElementNameFromBuildConfiguration(string expectedElementName)
+      {
+         if (expectedElementName.Equals(AppConstants.XmlNames.MoleculeBuildingBlock)) return AppConstants.XmlNames.Molecules;
+         if (expectedElementName.Equals(AppConstants.XmlNames.ObserverBuildingBlock)) return AppConstants.XmlNames.Observers;
+         if (expectedElementName.Equals(AppConstants.XmlNames.EventGroupBuildingBlock)) return AppConstants.XmlNames.EventGroups;
+         if (expectedElementName.Equals(AppConstants.XmlNames.MoBiReactionBuildingBlock)) return AppConstants.XmlNames.Reactions;
+         if (expectedElementName.Equals(AppConstants.XmlNames.ReactionBuildingBlock)) return AppConstants.XmlNames.Reactions;
+         if (expectedElementName.Equals(AppConstants.XmlNames.MoBiSpatialStructure)) return AppConstants.XmlNames.SpatialStructure;
+         if (expectedElementName.Equals(AppConstants.XmlNames.PassiveTransportBuildingBlock)) return AppConstants.XmlNames.PassiveTransports;
+         if (expectedElementName.Equals(AppConstants.XmlNames.ModelCoreSimulation)) return AppConstants.XmlNames.MoBiSimulation;
+
+         //this line is required if the file we are loading contains an old ModelCoreSimulation file
+         if (expectedElementName.Equals(AppConstants.XmlNames.MoBiSimulation)) return AppConstants.XmlNames.ModelCoreSimulation;
+         if (expectedElementName.Equals(AppConstants.XmlNames.CurveChart)) return AppConstants.XmlNames.Chart;
+
+         return expectedElementName;
+      }
+   }
+}
