+++ src/MoBi.Presentation/Serialization/SerializerRegister.cs
@@ -22,11 +22,6 @@ namespace MoBi.Presentation.Serialization
          coreSerializerRepository.AddPresentationSerializers();
          _register.PerformMappingForSerializerIn(container);
 
-         PerformMappingForMoBiSerializerRepository(container);
-      }
-
-      public void PerformMappingForMoBiSerializerRepository(IContainer container)
-      {
          var mobiSerializerRepository = container.Resolve<IMoBiXmlSerializerRepository>();
          mobiSerializerRepository.PerformMapping();
       }
