+++ src/MoBi.Presentation/Serialization/Xml/Serializer/MoBiXmlSerializerRepository.cs
@@ -1,35 +1,19 @@
-using MoBi.Core.Domain;
-using MoBi.Core.Serialization.Xml.Services;
-using OSPSuite.Core.Diagram;
-using OSPSuite.Core.Domain;
+using MoBi.Core.Serialization.Xml.Serializer;
 using OSPSuite.Core.Serialization.Xml;
 using OSPSuite.Presentation.Serialization;
 using OSPSuite.Presentation.Serialization.Extensions;
 using OSPSuite.Serializer;
-using OSPSuite.Serializer.Attributes;
 
 namespace MoBi.Presentation.Serialization.Xml.Serializer
 {
-   public class MoBiXmlSerializerRepository : OSPSuiteXmlSerializerRepository, IMoBiXmlSerializerRepository
+   public class MoBiXmlSerializerRepository : MoBiCoreXmlSerializerRepository
    {
       protected override void AddInitialSerializer()
       {
-         //To Add Serializer Classes from ModelCore
+         //Adds OSPSuite.Core + MoBi.Core serializers
          base.AddInitialSerializer();
 
-         //Remove serializer that will be overwritten in MoBi
-         RemoveSerializer("ReactionBuildingBlock");
-         RemoveSerializer("SpatialStructure");
-         RemoveSerializer("ModelCoreSimulation");
-
-         //To add IOSPSuiteXmlSerializer classes defined in MoBi
-         this.AddSerializers(x =>
-         {
-            x.Implementing<IOSPSuiteXmlSerializer>();
-            x.InAssemblyContainingType<IMoBiXmlSerializerRepository>();
-            x.UsingAttributeRepository(AttributeMapperRepository);
-         });
-
+         //Adds IOSPSuiteXmlSerializer classes defined in MoBi.Presentation
          this.AddSerializers(x =>
          {
             x.Implementing<IOSPSuiteXmlSerializer>();
@@ -37,7 +21,7 @@ namespace MoBi.Presentation.Serialization.Xml.Serializer
             x.UsingAttributeRepository(AttributeMapperRepository);
          });
 
-         //To add IPresentationXmlSerializer classes defined in MoBi
+         //Adds IPresentationXmlSerializer classes defined in MoBi.Presentation
          this.AddSerializers(x =>
          {
             x.Implementing<IPresentationXmlSerializer>();
@@ -46,14 +30,6 @@ namespace MoBi.Presentation.Serialization.Xml.Serializer
          });
 
          this.AddPresentationSerializers();
-      }
-
-      protected override void AddInitialMappers()
-      {
-         base.AddInitialMappers();
-         AttributeMapperRepository.AddAttributeMapper(new EnumAttributeMapper<NodeSize, SerializationContext>());
-         AttributeMapperRepository.AddAttributeMapper(new EnumAttributeMapper<NotificationType, SerializationContext>());
-         AttributeMapperRepository.AddAttributeMapper(new EnumAttributeMapper<OriginalQuantityValue.Types, SerializationContext>());
       }
    }
 }
