+++ src/MoBi.R/MoBi.R.csproj
@@ -1,29 +1,22 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <!--
-      Temporary: MoBi.R is meant to be cross-platform net10.0, but transitively pulls
-      OSPSuite.Presentation (Windows-only) via MoBi.CLI.Core / MoBi.Infrastructure /
-      MoBi.Presentation. Once that chain is broken, this TFM returns to net10.0 and the
-      assembly loads honestly on Linux / macOS .NET 10.
-      Tracked in https://github.com/Open-Systems-Pharmacology/MoBi/issues/2388.
-    -->
-    <TargetFramework>net10.0-windows</TargetFramework>
+    <TargetFramework>net10.0</TargetFramework>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
+    <Version Condition="'$(Version)' == ''">13.0.0</Version>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="OSPSuite.Core" Version="13.0.111" />
-    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.111" />
-    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.111" />
-    <PackageReference Include="OSPSuite.R" Version="13.0.111" />
+    <PackageReference Include="OSPSuite.Core" Version="13.0.113" />
+    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.113" />
+    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.113" />
+    <PackageReference Include="OSPSuite.R" Version="13.0.113" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\MoBi.Assets\MoBi.Assets.csproj" />
     <ProjectReference Include="..\MoBi.CLI.Core\MoBi.CLI.Core.csproj" />
     <ProjectReference Include="..\MoBi.Core\MoBi.Core.csproj" />
-    <ProjectReference Include="..\MoBi.Presentation\MoBi.Presentation.csproj" />
   </ItemGroup>
 
   <ItemGroup>
