+++ src/MoBi.R/Services/PassiveTransportsTask.cs
@@ -1,5 +1,5 @@
 using OSPSuite.Core.Domain.Builder;
-using ISerializationTask = MoBi.Presentation.Tasks.ISerializationTask;
+using ISerializationTask = MoBi.Core.Serialization.Services.ICoreSerializationTask;
 
 namespace MoBi.R.Services;
