+++ src/MoBi.R/Services/PathAndValuesBuildingBlockTask.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using MoBi.Assets;
 using OSPSuite.Core.Domain.Builder;
-using ISerializationTask = MoBi.Presentation.Tasks.ISerializationTask;
+using ISerializationTask = MoBi.Core.Serialization.Services.ICoreSerializationTask;
 
 namespace MoBi.R.Services;
