+++ src/MoBi.R/Services/ReactionsTask.cs
@@ -1,6 +1,6 @@
 using MoBi.Core.Domain.Model;
 using OSPSuite.Core.Domain.Builder;
-using ISerializationTask = MoBi.Presentation.Tasks.ISerializationTask;
+using ISerializationTask = MoBi.Core.Serialization.Services.ICoreSerializationTask;
 
 namespace MoBi.R.Services;
