+++ src/MoBi.UI/Services/ApplicationStartup.cs
@@ -11,8 +11,8 @@ using MoBi.Core.Extensions;
 using MoBi.Core.Services;
 using MoBi.Presentation;
 using MoBi.Presentation.Presenter;
+using MoBi.Core.Serialization.Xml;
 using MoBi.Presentation.Serialization;
-using MoBi.Presentation.Serialization.Xml;
 using MoBi.Presentation.Settings;
 using MoBi.Presentation.Views;
 using MoBi.UI.Diagram;
